Rename package application to service
