+++ service/builder/testdata/receivers_disabled.yaml
