+++ service/builder/testdata/receivers_enabled.yaml
