+++ service/factories_registration.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package application
+package service
 
 import (
 	// This is a temporary workaround to register all factories that are already
