+++ service/testdata/otelsvc-config.yaml
